Fix form macro with nette/application ≥ 3.2.2
Since https://github.com/nette/application/commit/1aba9da661b7b198c222d6bab6e69f4b0ef3879b,
`TemplateFactory` no longer registers the `FormsExtension`,
which caused the `form` macro to be missing.

Let’s re-add it in our `LatteFactory`.

The `CacheExtension`, which was also removed, is not necessary or
even detrimental for tests so we are going to ignore it.

With older `nette/application` versions, the `FormsExtension`
will be added a second time so the macros from the first
registration will be shadowed but that does not matter since
it is the same extension.